Refactor Ufs stream tests
