+++ core/client/fs/src/test/java/alluxio/client/file/ufs/UfsFileInStreamTest.java
@@ -24,10 +24,6 @@ import alluxio.client.file.FileSystemContext;
 import alluxio.client.file.URIStatus;
 import alluxio.client.file.options.FileSystemOptions;
 import alluxio.client.file.options.UfsFileSystemOptions;
-import alluxio.conf.Configuration;
-import alluxio.conf.InstancedConfiguration;
-import alluxio.conf.PropertyKey;
-import alluxio.conf.Source;
 import alluxio.exception.AlluxioException;
 import alluxio.grpc.DeletePOptions;
 import alluxio.underfs.UnderFileSystemFactoryRegistry;
@@ -44,36 +40,21 @@ import org.junit.runners.Parameterized;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.nio.ByteBuffer;
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.Optional;
 import java.util.Random;
-import java.util.UUID;
 
 /**
  * Add unit tests for {@link UfsFileInStream}.
  */
 @RunWith(Parameterized.class)
-public class UfsFileInStreamTest {
-  private static final int CHUNK_SIZE = 100;
-  private final InstancedConfiguration mConf;
-  private AlluxioURI mRootUfs;
-  private FileSystem mFileSystem;
-
-  @Parameterized.Parameters
-  public static Collection<Object[]> data() {
-    return Arrays.asList(new Object[][] {{false}, {true}});
-  }
-
+public class UfsFileInStreamTest extends AbstractUfsStreamTest {
   /**
    * Runs {@link UfsFileInStreamTest} with different configuration combinations.
    *
    * @param localDataCacheEnabled whether local data cache is enabled
    */
   public UfsFileInStreamTest(boolean localDataCacheEnabled) {
-    mConf = Configuration.copyGlobal();
-    mConf.set(PropertyKey.USER_CLIENT_CACHE_ENABLED,
-        PropertyKey.USER_CLIENT_CACHE_ENABLED.formatValue(localDataCacheEnabled), Source.RUNTIME);
+    super(localDataCacheEnabled);
   }
 
   /**
@@ -511,10 +492,6 @@ public class UfsFileInStreamTest {
       inStream.read(ByteBuffer.allocate(len), 0, len);
       assertEquals(CHUNK_SIZE / 4 - 1, inStream.remaining());
     }
-  }
-
-  private AlluxioURI getUfsPath() {
-    return new AlluxioURI(mRootUfs, String.valueOf(UUID.randomUUID()), true);
   }
 
   private FileInStream getStream(AlluxioURI ufsPath) throws IOException, AlluxioException {
